requirements.txt
amqp==5.2.0
asgiref==3.8.1
astroid==3.2.2
async-timeout==4.0.3
billiard==4.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2024.2
dill==0.3.8
Django==5.0.6
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-environ==0.11.2
flake8==7.0.0
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
gunicorn==22.0.0
httplib2==0.22.0
idna==3.7
isort==5.13.2
kombu==5.3.7
mccabe==0.7.0
oauthlib==3.2.2
packaging==24.1
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.46
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pyflakes==3.2.0
pylint==3.2.2
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
tomlkit==0.12.5
typing_extensions==4.12.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
